requirements.txt
sapiopylib~=2025.2.6.243
sapiopycommons==34.0.2.24.12.1
python-dotenv